+++ src/api/deprecated.rs
@@ -196,10 +196,10 @@ impl Engine {
     /// This method will be removed in the next major version.
     #[deprecated(since = "1.9.1", note = "use `register_fn` instead")]
     #[inline(always)]
-    pub fn register_result_fn<A: 'static, const N: usize, const C: bool, R: Variant + Clone>(
+    pub fn register_result_fn<A: 'static, const N: usize, const X: bool, R: Variant + Clone>(
         &mut self,
         name: impl AsRef<str> + Into<Identifier>,
-        func: impl RegisterNativeFunction<A, N, C, R, true> + SendSync + 'static,
+        func: impl RegisterNativeFunction<A, N, X, R, true> + SendSync + 'static,
     ) -> &mut Self {
         self.register_fn(name, func)
     }
@@ -218,10 +218,10 @@ impl Engine {
     #[deprecated(since = "1.9.1", note = "use `register_get` instead")]
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
-    pub fn register_get_result<T: Variant + Clone, const C: bool, V: Variant + Clone>(
+    pub fn register_get_result<T: Variant + Clone, const X: bool, R: Variant + Clone>(
         &mut self,
         name: impl AsRef<str>,
-        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C, V, true> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, X, R, true> + SendSync + 'static,
     ) -> &mut Self {
         self.register_get(name, get_fn)
     }
@@ -238,10 +238,10 @@ impl Engine {
     #[deprecated(since = "1.9.1", note = "use `register_set` instead")]
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
-    pub fn register_set_result<T: Variant + Clone, V: Variant + Clone, const C: bool, S>(
+    pub fn register_set_result<T: Variant + Clone, V: Variant + Clone, const X: bool>(
         &mut self,
         name: impl AsRef<str>,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C, (), true> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, X, (), true> + SendSync + 'static,
     ) -> &mut Self {
         self.register_set(name, set_fn)
     }
@@ -262,12 +262,12 @@ impl Engine {
     #[inline(always)]
     pub fn register_indexer_get_result<
         T: Variant + Clone,
-        X: Variant + Clone,
-        V: Variant + Clone,
-        const C: bool,
+        IDX: Variant + Clone,
+        R: Variant + Clone,
+        const X: bool,
     >(
         &mut self,
-        get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C, V, true> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>, IDX), 2, X, R, true> + SendSync + 'static,
     ) -> &mut Self {
         self.register_indexer_get(get_fn)
     }
@@ -286,12 +286,12 @@ impl Engine {
     #[inline(always)]
     pub fn register_indexer_set_result<
         T: Variant + Clone,
-        X: Variant + Clone,
-        V: Variant + Clone,
-        const C: bool,
+        IDX: Variant + Clone,
+        R: Variant + Clone,
+        const X: bool,
     >(
         &mut self,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C, (), true> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, IDX, R), 3, X, (), true> + SendSync + 'static,
     ) -> &mut Self {
         self.register_indexer_set(set_fn)
     }
@@ -551,7 +551,7 @@ impl Position {
 }
 
 #[allow(deprecated)]
-impl<'a, T: Variant + Clone> TypeBuilder<'a, T> {
+impl<T: Variant + Clone> TypeBuilder<'_, '_, T> {
     /// Register a custom fallible function.
     ///
     /// # Deprecated
@@ -562,15 +562,15 @@ impl<'a, T: Variant + Clone> TypeBuilder<'a, T> {
     /// This method will be removed in the next major version.
     #[deprecated(since = "1.9.1", note = "use `with_fn` instead")]
     #[inline(always)]
-    pub fn with_result_fn<S, A: 'static, const N: usize, const C: bool, R, F>(
+    pub fn with_result_fn<S, A: 'static, const N: usize, const X: bool, R, FUNC>(
         &mut self,
         name: S,
-        method: F,
+        method: FUNC,
     ) -> &mut Self
     where
         S: AsRef<str> + Into<Identifier>,
         R: Variant + Clone,
-        F: RegisterNativeFunction<A, N, C, R, true> + SendSync + 'static,
+        FUNC: RegisterNativeFunction<A, N, X, R, true> + SendSync + 'static,
     {
         self.with_fn(name, method)
     }
@@ -590,10 +590,10 @@ impl<'a, T: Variant + Clone> TypeBuilder<'a, T> {
     #[deprecated(since = "1.9.1", note = "use `with_get` instead")]
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
-    pub fn with_get_result<V: Variant + Clone, const C: bool>(
+    pub fn with_get_result<const X: bool, R: Variant + Clone>(
         &mut self,
         name: impl AsRef<str>,
-        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C, V, true> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, X, R, true> + SendSync + 'static,
     ) -> &mut Self {
         self.with_get(name, get_fn)
     }
@@ -611,10 +611,10 @@ impl<'a, T: Variant + Clone> TypeBuilder<'a, T> {
     #[deprecated(since = "1.9.1", note = "use `with_set` instead")]
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
-    pub fn with_set_result<V: Variant + Clone, const C: bool>(
+    pub fn with_set_result<const X: bool, R: Variant + Clone>(
         &mut self,
         name: impl AsRef<str>,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C, (), true> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, R), 2, X, (), true> + SendSync + 'static,
     ) -> &mut Self {
         self.with_set(name, set_fn)
     }
@@ -634,9 +634,9 @@ impl<'a, T: Variant + Clone> TypeBuilder<'a, T> {
     #[deprecated(since = "1.9.1", note = "use `with_indexer_get` instead")]
     #[cfg(any(not(feature = "no_index"), not(feature = "no_object")))]
     #[inline(always)]
-    pub fn with_indexer_get_result<X: Variant + Clone, V: Variant + Clone, const C: bool>(
+    pub fn with_indexer_get_result<IDX: Variant + Clone, R: Variant + Clone, const X: bool>(
         &mut self,
-        get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C, V, true> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>, IDX), 2, X, R, true> + SendSync + 'static,
     ) -> &mut Self {
         self.with_indexer_get(get_fn)
     }
@@ -654,9 +654,9 @@ impl<'a, T: Variant + Clone> TypeBuilder<'a, T> {
     #[deprecated(since = "1.9.1", note = "use `with_indexer_set` instead")]
     #[cfg(any(not(feature = "no_index"), not(feature = "no_object")))]
     #[inline(always)]
-    pub fn with_indexer_set_result<X: Variant + Clone, V: Variant + Clone, const C: bool>(
+    pub fn with_indexer_set_result<IDX: Variant + Clone, R: Variant + Clone, const X: bool>(
         &mut self,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C, (), true> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, IDX, R), 3, X, (), true> + SendSync + 'static,
     ) -> &mut Self {
         self.with_indexer_set(set_fn)
     }