+++ src/api/register.rs
@@ -64,26 +64,26 @@ impl Engine {
     pub fn register_fn<
         A: 'static,
         const N: usize,
-        const C: bool,
+        const X: bool,
         R: Variant + Clone,
-        const L: bool,
-        F: RegisterNativeFunction<A, N, C, R, L> + SendSync + 'static,
+        const F: bool,
+        FUNC: RegisterNativeFunction<A, N, X, R, F> + SendSync + 'static,
     >(
         &mut self,
         name: impl AsRef<str> + Into<Identifier>,
-        func: F,
+        func: FUNC,
     ) -> &mut Self {
-        let param_types = F::param_types();
+        let param_types = FUNC::param_types();
 
         #[cfg(feature = "metadata")]
-        let mut param_type_names = F::param_names()
+        let mut param_type_names = FUNC::param_names()
             .iter()
             .map(|ty| format!("_: {}", self.format_type_name(ty)))
             .collect::<crate::FnArgsVec<_>>();
 
         #[cfg(feature = "metadata")]
-        if F::return_type() != TypeId::of::<()>() {
-            param_type_names.push(self.format_type_name(F::return_type_name()).into());
+        if FUNC::return_type() != TypeId::of::<()>() {
+            param_type_names.push(self.format_type_name(FUNC::return_type_name()).into());
         }
 
         #[cfg(feature = "metadata")]
@@ -96,10 +96,10 @@ impl Engine {
         let is_pure = true;
 
         #[cfg(any(not(feature = "no_index"), not(feature = "no_object")))]
-        let is_pure = is_pure && (F::num_params() != 3 || fn_name != crate::engine::FN_IDX_SET);
+        let is_pure = is_pure && (FUNC::num_params() != 3 || fn_name != crate::engine::FN_IDX_SET);
         #[cfg(not(feature = "no_object"))]
         let is_pure =
-            is_pure && (F::num_params() != 2 || !fn_name.starts_with(crate::engine::FN_SET));
+            is_pure && (FUNC::num_params() != 2 || !fn_name.starts_with(crate::engine::FN_SET));
 
         let func = func.into_callable_function(is_pure, true);
 
@@ -278,12 +278,12 @@ impl Engine {
     /// Register a fallible type iterator for an iterable type with the [`Engine`].
     /// This is an advanced API.
     #[inline(always)]
-    pub fn register_iterator_result<T, X>(&mut self) -> &mut Self
+    pub fn register_iterator_result<T, R>(&mut self) -> &mut Self
     where
-        T: Variant + Clone + IntoIterator<Item = RhaiResultOf<X>>,
-        X: Variant + Clone,
+        T: Variant + Clone + IntoIterator<Item = RhaiResultOf<R>>,
+        R: Variant + Clone,
     {
-        self.global_namespace_mut().set_iterable_result::<T, X>();
+        self.global_namespace_mut().set_iterable_result::<T, R>();
         self
     }
     /// Register a getter function for a member of a registered type with the [`Engine`].
@@ -328,10 +328,10 @@ impl Engine {
     /// ```
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
-    pub fn register_get<T: Variant + Clone, const C: bool, V: Variant + Clone, const L: bool>(
+    pub fn register_get<T: Variant + Clone, const X: bool, R: Variant + Clone, const F: bool>(
         &mut self,
         name: impl AsRef<str>,
-        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C, V, L> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, X, R, F> + SendSync + 'static,
     ) -> &mut Self {
         self.register_fn(crate::engine::make_getter(name.as_ref()), get_fn)
     }
@@ -378,10 +378,10 @@ impl Engine {
     /// ```
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
-    pub fn register_set<T: Variant + Clone, const C: bool, V: Variant + Clone, const L: bool>(
+    pub fn register_set<T: Variant + Clone, const X: bool, R: Variant + Clone, const F: bool>(
         &mut self,
         name: impl AsRef<str>,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C, (), L> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, R), 2, X, (), F> + SendSync + 'static,
     ) -> &mut Self {
         self.register_fn(crate::engine::make_setter(name.as_ref()), set_fn)
     }
@@ -434,16 +434,16 @@ impl Engine {
     #[inline(always)]
     pub fn register_get_set<
         T: Variant + Clone,
-        const C1: bool,
-        const C2: bool,
-        V: Variant + Clone,
-        const L1: bool,
-        const L2: bool,
+        const X1: bool,
+        const X2: bool,
+        R: Variant + Clone,
+        const F1: bool,
+        const F2: bool,
     >(
         &mut self,
         name: impl AsRef<str>,
-        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C1, V, L1> + SendSync + 'static,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C2, (), L2> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, X1, R, F1> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, R), 2, X2, (), F2> + SendSync + 'static,
     ) -> &mut Self {
         self.register_get(&name, get_fn).register_set(&name, set_fn)
     }
@@ -500,13 +500,13 @@ impl Engine {
     #[inline]
     pub fn register_indexer_get<
         T: Variant + Clone,
-        X: Variant + Clone,
-        const C: bool,
-        V: Variant + Clone,
-        const L: bool,
+        IDX: Variant + Clone,
+        const X: bool,
+        R: Variant + Clone,
+        const F: bool,
     >(
         &mut self,
-        get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C, V, L> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>, IDX), 2, X, R, F> + SendSync + 'static,
     ) -> &mut Self {
         #[cfg(not(feature = "no_index"))]
         assert!(
@@ -587,13 +587,13 @@ impl Engine {
     #[inline]
     pub fn register_indexer_set<
         T: Variant + Clone,
-        X: Variant + Clone,
-        const C: bool,
-        V: Variant + Clone,
-        const L: bool,
+        IDX: Variant + Clone,
+        const X: bool,
+        R: Variant + Clone,
+        const F: bool,
     >(
         &mut self,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C, (), L> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, IDX, R), 3, X, (), F> + SendSync + 'static,
     ) -> &mut Self {
         #[cfg(not(feature = "no_index"))]
         assert!(
@@ -675,16 +675,16 @@ impl Engine {
     #[inline(always)]
     pub fn register_indexer_get_set<
         T: Variant + Clone,
-        X: Variant + Clone,
-        const C1: bool,
-        const C2: bool,
-        V: Variant + Clone,
-        const L1: bool,
-        const L2: bool,
+        IDX: Variant + Clone,
+        const X1: bool,
+        const X2: bool,
+        R: Variant + Clone,
+        const F1: bool,
+        const F2: bool,
     >(
         &mut self,
-        get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C1, V, L1> + SendSync + 'static,
-        set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C2, (), L2> + SendSync + 'static,
+        get_fn: impl RegisterNativeFunction<(Mut<T>, IDX), 2, X1, R, F1> + SendSync + 'static,
+        set_fn: impl RegisterNativeFunction<(Mut<T>, IDX, R), 3, X2, (), F2> + SendSync + 'static,
     ) -> &mut Self {
         self.register_indexer_get(get_fn)
             .register_indexer_set(set_fn)